feat: [mergeability-follow-up] - [Reduce provider package file count]
