+++ docs/internal/processes/api-relevant-files.md
@@ -7,6 +7,7 @@ Use this map when changing the public REST contract.
 - Provider-specific response-stream normalization must stay internal. When changing Cursor/Codex live-stream parsers, extend `pkg/api/contracttests/openapi_contract_surface_test.go` and `pkg/api/contracttests/generated_contract_events_test.go` with behavioral boundary assertions on published schemas and generated event JSON; do not use artifact string scans as the proof that parser-only fields stay internal.
 - Live provider progress publication into internal session streams belongs in `pkg/workers/provider/inference_provider.go` (`WithInferenceProgressPublisher`, provider terminal success/failure markers, subprocess fragment kinds), `pkg/service/factory_build.go` (inject the session-owned publisher into ScriptWrapProvider construction and default subprocess runner wrapping), and `pkg/service/runtime_sessions.go` (`mapInferenceProgressFragment`, `inferenceProgressPublisher`, response-stream metrics/logging). Keep provider progress internal-only: normalize at the provider boundary, publish only `SessionResponseStreamEvent` kinds, and do not append raw provider progress into canonical `FactoryEvent` history or public API schemas. Treat publication as additive and failure-safe at the session-owned publisher seam: recover publication panics, emit bounded `session_response_stream.degraded` metrics and warn logs with reason labels such as `STREAM_UNAVAILABLE` or `PUBLISH_PANIC`, and never let internal stream faults replace the provider's final `WorkResult`. Additive verification belongs in `pkg/service/testmain_test.go`, where model-worker execution should prove an injected internal progress publisher can run alongside canonical inference-event recording without changing the final `WorkResult` or public event history; pressure and lagging-consumer verification belongs in `pkg/service/runtime_session_runtime_test.go`, where the live-session publisher seam should prove no-subscriber and slow-consumer cases degrade through bounded retained-window compaction rather than blocking execution.
 - `pkg/service` is under `pkg-maint` enforcement for cyclomatic complexity and function-line limits even in tests. When extending the internal provider/session-stream verification path in `pkg/service/factory_build.go`, `pkg/service/testmain_test.go`, or `pkg/service/runtime_session_runtime_test.go`, prefer small fixture/assertion helpers over growing constructor branches or long same-package test bodies so mergeability fixes do not become a recurring CI blocker.
+- `pkg/workers/provider` is also under a hard `make pkg-file-count` cap of 15 owned Go files for the package directory. Provider-stream follow-ups should prefer merging tiny focused helpers/tests into existing provider files before adding new companion files, or remove/recombine older small files in the same change, so mergeability fixes do not replace one CI blocker with another.
 
 - `api/openapi-main.yaml` is the authored OpenAPI entrypoint. Register new component fragments under `components.schemas` here.
 - `api/components/schemas/api/` contains API request/response fragments.
