+++ pkg/workers/provider/provider_errors_test.go
@@ -9,6 +9,51 @@ import (
 	"github.com/portpowered/infinite-you/pkg/logging"
 )
 
+func loadProviderErrorCorpusForTest(t *testing.T) ProviderErrorCorpus {
+	t.Helper()
+
+	corpus, err := LoadProviderErrorCorpus()
+	if err != nil {
+		t.Fatalf("LoadProviderErrorCorpus() error = %v", err)
+	}
+	return corpus
+}
+
+func providerErrorCorpusEntryForTest(t *testing.T, name string) ProviderErrorCorpusEntry {
+	t.Helper()
+
+	entry, ok := loadProviderErrorCorpusForTest(t).Entry(name)
+	if !ok {
+		t.Fatalf("provider error corpus entry %q not found", name)
+	}
+	return entry
+}
+
+func providerErrorCorpusEntryLabel(entry ProviderErrorCorpusEntry) string {
+	if entry.UpstreamSourceCase == "" {
+		return entry.Name
+	}
+	return entry.Name + " [" + entry.UpstreamSourceCase + "]"
+}
+
+func providerErrorCorpusLastErrorLine(t *testing.T, entry ProviderErrorCorpusEntry) string {
+	t.Helper()
+
+	var lastMatch string
+	for _, stream := range []string{entry.Stderr, entry.Stdout} {
+		for _, line := range strings.Split(stream, "\n") {
+			trimmed := strings.TrimSpace(line)
+			if strings.HasPrefix(trimmed, "ERROR:") {
+				lastMatch = trimmed
+			}
+		}
+	}
+	if lastMatch == "" {
+		t.Fatalf("provider error corpus entry %q contains no ERROR: line", providerErrorCorpusEntryLabel(entry))
+	}
+	return lastMatch
+}
+
 func TestNewProviderError_AssignsDeterministicFamilyFromType(t *testing.T) {
 	testCases := []struct {
 		name       string
